refactor(cli/zk): move printing classes into their own file
This sets the stage to the cli bits testable.
